Don't use CreateChildAtOffset to create synthetic child values.
We were using child.CreateChildAtOffset(name, 0, type) to create a child
value starting at offset 0 of the child object. That is, starting at
(&child + offset).

However, when child is of pointer type, child.CreateChildAtOffset
will create a child from the memory at (child + offset), dereferencing
the pointer. See the included test for an example of incorrect output
caused by this.

This patch changes it to use CreateValueFromData instead.
child.sbvalue().GetData() preserves the value of the original pointer,
which is what we want.