Make gdb.lookup('char') work for binaries built with `-funsigned-char`.
When the target binary was built with `-funsigned char`, lldb treats
'char' as an alias for 'unsigned char' and this causes type lookups for
'char' to fail. As a workaround, we detect this special case in
`gdb.lookup_type` and return `gdb.lookup_type('unsigned char') in case
of lookup failure.