Use CreateValueFromAddress to create child values when possible.
When we return an SBValue for a synthetic child, we create a new one.
Using CreateValueFromData works, but the resulting SBValue won't have
a memory address, so if another prettyprinter down the road tries to
access the address of the value to print, it will fail.

Ideally we'd want to return exactly the same SBValue, but with a
different name. Because the SB API doesn't offer that option (see also
https://reviews.llvm.org/D31371), the next best thing I've found is to
use CreateValueFromAddress when possible, so the synthetic SBValue will
have an address.