Add workaround for casting integers to a larger integer type.
`SBValue::Cast` does essentially a reinterpret_cast of the address of
the value. When casting from an integer type to a larger integer type,
this results in garbage from adjacent positions in memory.

This change adds logic to detect this case and create a new result value
of the target type that contains the original numeric value. This new
value is not backed by any actual program memory (`result.address` is
`None`), but this matches gdb behavior.