Add a way to specify a base path for .debug_gdb_scripts during autoload.
Right now, the autoload script will use the paths in .debug_gdb_scripts
as-is when trying to open the script files. That means that relative
paths will be interpreted as relative to the current working directory
at script load time.

With this change it's possible to have some script that searches, for
example, the root of the current repository and passes it to autoload
initialization so scripts can be found regardless of what exact
subdirectory we're in.