Make sure `num_children` in GALA synthetic child provider returns int.
lldb used to be happy when `len(self._children) / 2` returned a float
because of the division. But some tests have started failing in recent
builds of lldb because of this. I suspect the reason is a difference in
the Python version, as I've seen it work on an lldb with Python 3.9 and
fail on an lldb with Python 3.10.

num_children should always return an int anyway, so change the
expression to use the integer division operator.