Use SBData.GetByteSize instead of SBSection.GetFileByteSize.
If .debug_gdb_scripts is compressed SBSection.GetFileByteSize will
return the size of the compressed byte stream, but what we actually want
is the size of the uncompressed data.

So instead we call SBSection.GetSectionData to get an SBData object for
the full uncompressed section, and ask that object for the data size.