Avoid casting SBValues in gdb.Value.__getitem__.
The GALA implementation of gdb.Value.__getitem__ used to `Cast` an
SBValue to its typedef-stripped type before trying to access its
members. This caused the following sequence to fail in a surprising way:

1. a prettyprinter caches the result of an expensive gdb.parse_and_eval
   call.
2. The user restarts the process.
3. Another instance of the prettyprinter tries to evaluate
   `cached_value["member"]`, and it gets a Value object for the member
   that doesn't actually contain the value of the member and will print
   as just an empty string.

It seems we can do away with the Cast (all tests pass), so I've removed
it and made a few little changes to __getitem__ to make it look more
coherent.