Make prettyprinters follow pointers.
lldb automatically dereferences pointers when asking for children of a
pointer SBValue. This causes problems when using lldb-vscode. Expanding
a pointer variable in lldb-vscode will retrieve the children of a
pointer SBValue, and with "skip pointers" this always results in
non-synthetic children. The pointee object does never appear as part of
the chain of expanded values, so the prettyprinter never kicks in.

So, in order for GALA prettyprinters to behave correctly with
lldb-vscode, like built-in lldb formatters do, we need to switch to
follow pointers even if this doesn't match gdb behavior.

Because prettyprinters written for gdb won't necessarily work correctly
when getting a pointer instead of an object, this change also adds
a couple of calls to SBValue.Dereference() in the wrappers for summaries
and synthetic child providers before jumping to the original code.