Reduce console spam in case of prettyprinter errors.
Printing a prettyprinter exception in response to a user `print` command
in the CLI is not problematic. Inside an IDE, however, lldb-vscode will
try to print all local variables every time the program stops (including
every time the user manually steps through the code).

This can result in tons of log spam in the console window for relatively
innocent cases. For example, when local variables are not yet
initialized, or not yet in scope, prettyprinters can fail with a memory
read error trying to follow an invalid pointer.

With this change, when we catch an exception in a gdb prettyprinter we
include information about the error as a part of the generated summary
string or the synthetic children. This way it will appear next to the
problematic variable, which makes it easier for the user to understand
what happened, and no longer clutters the console window.

The drawback is that when a child provider fails, we'll generate a
fictional first child to hold the error message, which is displayed
together with the actual members of the value. Hopefully a child at the
top named "LLDB ERROR" won't be confusing for users.