Add value name on error message when calling children().
When using DAP, the IDE prints all local variables all the time. This
increases the likelihood of printing an uninitialized variable, and also
makes it hard to know which specific variable caused a prettyprinter
error dumped to the console window.

Make a best-effort attempt to print the variable name in case of error
so it's easier to find out which one caused the problem.