Short-circuit lookup_type for builtin types (#14)
Use the existing BUILTIN_TYPE_NAME_TO_BASIC_TYPE to quickly return the
corresponding builtin type. FindTypes("char") can take a fair amount of
time, particularly if the pretty printer does it in a loop, and this can
result in significant speedups.

It also allows us to remove the `unsigned char` hackaround.