Fix print_elements test
The current implementation of gdb.Value.__str__ doesn't print children,
and the test wasn't checking that the prettyprinter yielded any
children, so it wasn't really testing anything.

Use the `p` command instead from the test as a workaround (so children
get printed), and add a positive CHECK to verify that the printer yields
as many elements as desired, but not more.