Use SBType::FindFirstType for gdb.lookup_type (#18)
.. instead of `FindTypes(basename)` + custom filtering.

This makes a huge difference when it comes to performance, because
something like `FindTypes("iterator")` forces lldb to return all
iterator types anywhere in the program. FindFirstType bails out the
search as soon as it finds the first match (which is all that we're
interested in).

It's unclear why the function was implemented as it was, but a part of
the reason was likely the fact that FindTypes had problems with looking
up/parsing compex type names containing templates. This has now been
(mostly) fixed in llvm/llvm-project#91586, and appears that the
simplified version performs satisfactorily.