Add a pretty printer extension to directly return the number of children
This avoids us needing to enumerate the children one by one (if the
pretty printer supports the extension).

To keep the pretty printer interface consistent, I've define the
extensions in terms of the number of children returned/yielded by the
`children()` function rather than the number of entries in a map, even
though this means that most map pretty printers will need to multiple
the number of elemnts by two, only for GALA to divide it again.

I'm also passing down the `max_count` value we get from lldb, in it's
useful in some pretty printers for limiting the amount of work.

With llvm/llvm-project#93946, lldb will pass down reasonable values for
max_count, so we might be able to remove the limiting hack on the
fallback path, but I'm leaving that for a separate patch/discussion.