Update gdb.Value.__str__ to check synthetic values.
This udpates gdb.Value.__str__ to iterate through, and return,
synthetic childre values if there are any, when appropriate.  It
also fixes a small bug in print_elements.test (the assumed return
values for negative numbers is no longer correct), and it updates
that test to test this new functionality.